+++ src/test/java/gift/product/application/dto/request/ProductRequestTest.java
@@ -1,9 +1,8 @@
-package gift.controller.dto.request;
+package gift.product.application.dto.request;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import gift.common.validation.ValidateErrorMessage;
-import gift.product.application.dto.request.ProductRequest;
 import jakarta.validation.ConstraintViolation;
 import jakarta.validation.Validation;
 import jakarta.validation.Validator;
